+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,14 +2,14 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_133299: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_133177: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_131640: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_131630: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_133298: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_128612: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_128591: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_133300: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_133351: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_133229: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_128664: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_128643: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_133352: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_131692: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_131682: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_133350: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;