+++ src/music/mod.rs
@@ -1,5 +1,5 @@
 mod plugin;
-pub mod resources;
+mod resources;
 mod systems;
 
 pub use plugin::MusicPlugin;